refactor(components): react select input
