+++ src/lib/components/forms/asset-input/AssetInputOption.tsx
@@ -2,11 +2,14 @@ import { Box, Flex, Grid, Text } from "@chakra-ui/react";
 import type { OptionProps } from "chakra-react-select";
 import { components } from "chakra-react-select";
 
+import type { SelectInputOption } from "../SelectInput";
 import { TokenImageRenderWithCache } from "lib/components/token";
-import type { AssetOption } from "lib/types";
+import type { AssetOptionValue } from "lib/types";
 import { getTokenLabel } from "lib/utils";
 
-export const AssetInputOption = (props: OptionProps<AssetOption>) => {
+export const AssetInputOption = (
+  props: OptionProps<SelectInputOption<AssetOptionValue>>
+) => {
   const { data, isSelected } = props;
   const textMainColor =
     data.isDisabled && !isSelected ? "text.disabled" : "text.main";
@@ -38,27 +41,29 @@ export const AssetInputOption = (props: OptionProps<AssetOption>) => {
         )}
         <Flex justifyContent="space-between" alignItems="center">
           <Flex gap={2} alignItems="center">
-            <Box opacity={data.isDisabled && !isSelected ? 0.6 : 1}>
-              <TokenImageRenderWithCache
-                src={String(data.logo)}
-                alt={getTokenLabel(String(data.id), data.label)}
-                width={24}
-                height={24}
-              />
-            </Box>
+            {!!data.value.logo && (
+              <Box opacity={data.isDisabled && !isSelected ? 0.6 : 1}>
+                <TokenImageRenderWithCache
+                  src={data.value.logo}
+                  alt={getTokenLabel(data.value.id, data.label)}
+                  width={24}
+                  height={24}
+                />
+              </Box>
+            )}
             <Text variant="body2" color={textMainColor}>
               {data.label}
             </Text>
           </Flex>
           <Flex direction="column" alignItems="flex-end">
             <Text variant="body2" color={textMainColor}>
-              {data.formatted || "0.000000"}
+              {data.value.formatted || "0.000000"}
             </Text>
             <Text
               variant="body3"
               color={data.isDisabled ? "text.disabled" : "text.dark"}
             >
-              {`(${data.price || "$0.00"})`}
+              {`(${data.value.price || "$0.00"})`}
             </Text>
           </Flex>
         </Flex>
