+++ src/lib/components/modal/wasm-verify-submit/WasmVerifySubmitFormSelectNoOptionsMessage.tsx
@@ -1,11 +1,13 @@
 import { Text } from "@chakra-ui/react";
-import { components } from "chakra-react-select";
 import type { NoticeProps } from "chakra-react-select";
+import { components } from "chakra-react-select";
 
-import type { WasmVerifySubmitFormOption } from "./WasmVerifySubmitFormSelectOption";
+import type { SelectInputOption } from "lib/components/forms";
+
+import type { WasmVerifySubmitFormOptionValue } from "./WasmVerifySubmitFormSelectOption";
 
 export const WasmVerifySubmitFormSelectNoOptionsMessage = (
-  props: NoticeProps<WasmVerifySubmitFormOption>
+  props: NoticeProps<SelectInputOption<WasmVerifySubmitFormOptionValue>>
 ) => (
   <components.NoOptionsMessage {...props}>
     <Text px={3} textAlign="left" variant="body2">
